Changed ABCs for TS and TSD
