+++ src/timeatlas/abstract/__init__.py
@@ -1,24 +1,22 @@
-from .abstract_analysis import AbstractAnalysis
 from .abstract_base_detector import AbstractBaseDetector
+from .abstract_base_generator import AbstractBaseGenerator
 from .abstract_base_model import AbstractBaseModel
+from .abstract_base_time_series import AbstractBaseTimeSeries
 from .abstract_io import (
     AbstractInput,
     AbstractOutputJson,
     AbstractOutputPickle,
     AbstractOutputText,
 )
-from .abstract_processing import AbstractProcessing
-from .abstract_base_generator import AbstractBaseGenerator
 
 __all__ = [
-    "AbstractAnalysis",
     "AbstractBaseDetector",
+    "AbstractBaseGenerator",
     "AbstractBaseModel",
+    "AbstractBaseTimeSeries",
+    # Abstract IO
     "AbstractInput",
     "AbstractOutputJson",
     "AbstractOutputPickle",
     "AbstractOutputText",
-    "AbstractProcessing",
-    "AbstractBaseGenerator",
-
 ]
